backend/requirements.txt
Django==6.0.4
channels==4.2.0
channels-redis==4.2.1
daphne==4.1.2
requests==2.32.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
django-filter==25.2
cryptography==44.0.0
redis==7.4.0
dj-database-url==2.3.0
gunicorn==23.0.0
psycopg[binary]==3.2.9
whitenoise==6.9.0
Pillow==11.0.0
google-auth==2.38.0
